requirements.txt
# Core dependencies
fastmcp==2.11.1
uv==0.1.0
python-dotenv==1.0.0
pydantic==2.5.0

# AI Provider SDKs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
mistralai==0.1.3

# Caching and similarity search
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.24.3

# ADK and MCP tools (for agent integration)
google-adk==0.2.0
mcp==1.0.0

# API framework
fastapi==0.109.0
uvicorn==0.27.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.1
flake8==6.1.0
